Create ArrayOfProducts
In this file there is an array that is being inserted and we want the output to have the same length as the original array however each index i has the product of all the the values from the original array except for the value of index i (from original array).   there are two solutions, one uses brute force and has a time complexity of O(n^2) and the other has a complexity of O(n).  The way the O(n) algorithm works is by creating two arrays (left and right), these two arrays have the product to the left of index i (left array) and the other has the products to the right of index i (right array) then afterwards both arrays are used to get the product of all the values in the original array except at index i.