SumOfTwo
This problem was to find if there are two numbers in a given array that have a sum of a given value k.  This was done by using HashSet.  My algorithm ensures that numbers are not duplicated and that the sum of two distinct have a sum of a given value k.  The algorithm returns true if there are two such distinct integers and also returns false if there are no distinct values that has a sum of k.